A slightly faster default x-product filter function, and hopefully not as prone to sensitivity to the grammar factorization. Removed casts of ints -> shorts and vice versa. We'll allow 2^15 entries on either side of grammar rules. That should accomodate all the enumerable grammars we care about.

+++ java/src/edu/ohsu/cslu/parser/CsrSpmvParser.java
@@ -116,7 +116,7 @@ public class CsrSpmvParser extends SparseMatrixVectorParser<CsrSparseMatrixGramm
 
                 for (int j = rightCell.offset(); j <= rightCell.maxRightChildIndex(); j++) {
 
-                    final int childPair = cpf.pack(leftChild, (short) nonTerminalIndices[j]);
+                    final int childPair = cpf.pack(leftChild, nonTerminalIndices[j]);
                     totalCartesianProductEntriesExamined++;
 
                     if (!cpf.isValid(childPair)) {
